wip - separating channel management into new module
